added server.js and controller for volunteer history
